+++ classes/BackyardBriefApiClient.php
@@ -24,13 +24,13 @@ class BackyardBriefApiClient
     /**
      *
      * @param string $apiUrl
-     * @param mixed $appLogFolder OPTIONAL string without trailing / or if null => the applogs will not be saved at all
+     * @param ?string $appLogFolder OPTIONAL string without trailing / or if null => the applogs will not be saved at all
      * @param \Psr\Log\LoggerInterface $logger OPTIONAL but really recommended
      */
     public function __construct($apiUrl, $appLogFolder = null, LoggerInterface $logger = null)
     {
         //error_log("debug: " . __CLASS__ . ' ' . __METHOD__);
-        $this->logger = $logger;
+        $this->logger = is_null($logger) ? new \Psr\Log\NullLogger() : $logger;
         $this->apiUrl = $apiUrl;
         $this->appLogFolder = $appLogFolder;
     }